requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==1.0.0
pytest==7.4.2